requirements.txt
django==5.2.1
djangorestframework==3.15.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-filter==24.1
django-jazzmin==2.7.0
Pillow==10.3.0
mysqlclient==2.2.2
django-debug-toolbar==4.3.0
python-dotenv==1.0.1
whitenoise==6.6.0
gunicorn==22.0.0 